Move ImageJ macro files out of Java source folder
For now, I put them in the resources folder, so that they get included
in the misc_plugins.jar that Maven builds. However, the macros are not
included in plugins.config (I am not even sure you can do this...), so
they will still not automatically appear in the menus.