VisualizeTiles: fix Z values and tile colors
The intended tile color was overwriting the Z depth, and the tile color
was never actually being set.

The tile color for each OME Image is now set to a constant-but-random
color, to easily differentiate between 5D Image blocks.

In the case of Prairie data, this visualization manifests as columns of
tiles from the same Image (since there are multiple Z positions at the
same X,Y stage position).