TilesToROIs: don't make a subPixel ROI
All ROIs in this plugin were being constructed with doubles, which
causes ImageJ to treat them as having sub-pixel accuracy (and convert
them to ints anyway).

So instead we do the casts to ints ahead of time and avoid the sub-pixel
accuracy flag being set. This allows the ROIs to be modified in a way
one would typically expect of a ROI.