TilesToROIs: display all slices
TilesToROIs was processing all tiles, but many of them - when added as
ROIs - were discarded for being duplicates, as their slice index was
never set.

Tiles are now mapped to slice indices. Then the entire image stack is
created, with each resulting Roi indexed under the correct slice.

This allows visualization of all tiles in the dataset, and not just the
first tile discovered in each position.