TileUtils: normalize ROI unit calibration
When computing coordinates in TileUtils, the height and width of each
tile (on a per-series basis) are normalized to the first valid x,y
calibration values in the dataset.

If series 2 had a calibration of 0.5um and series 0 had a calibration of
0.1um, then each roi from series 2 would have its height and width
multiplied by (0.5 / 0.1).

Stage positions are also converted to uncalibrated units (position /
calibration).

This allows us report calibrated values in consumers of TileUtils.
TilesToROIs now applies the first calibrated values to the ImagePlus it
creates.. so it should be clearer what the actual size of the displayed
regions are.