Update to Bio-Formats 5.1.0 release
This bumps the pom-loci parent POM to 2.0.0, which inherits
Bio-Formats 5.1.0 from the pom-scijava ancestor POM.

This makes the necessary code changes to account for backwards
incompatibilities in the Bio-Formats API from 5.0.x to 5.1.x.
In so doing, it cleans up the metadata handling just a little,
with some new utility methods in TileUtils. In theory,
non-micron units may work, though this is completely untested,
and also depends on how stage position units are specified.